requirements.txt
PyYAML>=5.3.1
numpy~=1.24.2
opencv-python~=4.7.0.72
